Modified main.py and robust division calculator1

+++ programming_paradigm/robust_division_calculator.py
@@ -1,7 +1,7 @@
 def safe_divide(numerator, denominator):
     try:
-        numerator = int(numerator)
-        denominator = int(denominator)
+        numerator = float(numerator)
+        denominator = float(denominator)
     except ValueError:
         print("Error: Please enter numeric values only.")
         return
